added missing if
git-svn-id: http://www.pimcore.org/svn/pimcore/private/core/trunk@4442 1f8fe7d8-47f0-464c-8d0a-336f4953ab05

+++ pimcore/modules/admin/controllers/ObjectController.php
@@ -741,7 +741,7 @@ class Admin_ObjectController extends Pimcore_Controller_Action_Admin
 
         } else if ($this->getParam("id")) {
             $object = Object_Abstract::getById($this->getParam("id"));
-            if ($object->isAllowed("delete")) {
+            if ($object && $object->isAllowed("delete")) {
                 $object->delete();
 
                 $this->_helper->json(array("success" => true));
